requirements.txt
audioread==3.0.1
filetype==1.2.0
Pillow==10.2.0
pyglet==1.5.28
python_magic==0.4.27
PyWavefront==1.3.3
Requests==2.31.0
validators==0.23.2
